pyproject.toml
[tool.poetry]
name = "at-config"
version = "1.0.0"
description = ""
authors = ["grigadnal625 <grigandal625@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
pydantic = "^2.10.6"
async-property = "^0.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pre-commit = "^4.1.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
